Extracting the creation of CallPattern from MocksRepository
- Extracting the logic for creating CallPattern from MocksRepository
- Rearanging the fields and properties to be in the begining of MocksRepository as they were scattered.
